refactor: split Android platform module into focused submodules
Break down the monolithic `src/platforms/android/index.ts` (~1533 LOC) into
focused submodules to improve maintainability and targeted code review.

New submodules:
- `adb.ts` (21 LOC) — shared helpers: adbArgs, ensureAdb, sleep, isClipboardShellUnsupported
- `app-lifecycle.ts` (507 LOC) — app resolve/list/open/close/install/reinstall/appstate + bundletool
- `input-actions.ts` (378 LOC) — press/swipe/longPress/type/fill/scroll/scrollIntoView
- `device-input-state.ts` (184 LOC) — keyboard state/dismiss + clipboard read/write
- `settings.ts` (314 LOC) — appearance/permission/fingerprint/wifi/airplane/location
- `notifications.ts` (61 LOC) — push notification broadcast
- `snapshot.ts` (98 LOC) — snapshot + screenshot + UI hierarchy dump
- `index.ts` (43 LOC) — thin re-export layer preserving all public exports

Closes #154
